requirements.txt
grpcio
grpcio-tools
boto3
pillow
python-dotenv